Alfa version

+++ config.sh
@@ -49,29 +49,66 @@ fi
 
 # 2. change default shell by zsh
 # ===============================
-printf "\nInstalling zsh\n"
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-
+printf "\nInstalling oh-my-zsh? [y/n] "; read OK
+if [ "$OK" != "Y" ] && [ "$OK" != "y" ]
+then
+  printf "\n Exiting installation...\n"
+  exit 0
+else
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+fi
 
 # 3. homebrew
 # ===============================
 if ! command -v brew > /dev/null 2>&1; then
   printf "\nInstalling Homebrew\n"
   ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+else
+  printf "\nHomebrew Installed\n"
 fi
 
 # 3.1. update homebrew
 # -------------------------------
-brew update
+printf "\nUpdating Homebrew\n"
+brew update 
 brew update
 brew upgrade
 
 # 3.2. install some taps
 # -------------------------------
-brew install vim neovim tmux z fortune cowsay coreutiles \
-             imagemagick pandoc markdown syncthing htop  \
-             ffmpeg reattach-to-user-namespace m4 ranger \
-             node git wget 
+taps=(
+  vim
+  neovim
+  tmux
+  z
+  fortune
+  cowsay
+  ranger
+  coreutiles
+  imagemagick
+  pandoc
+  markdown
+  syncthing
+  htop
+  ffmpeg
+  reattach-to-user-namespace
+  m4
+  python
+  node
+  git
+  wget 
+  )
+
+printf "\nThe following taps will be installed\n"
+printf '   - %s\n' "${taps[@]}"
+printf "\nContinue [y/n] "; read OK
+if [ "$OK" != "Y" ] && [ "$OK" != "y" ]
+then
+  printf "\n Exiting installation...\n"
+  exit 0
+else
+  brew install ${taps[@]}
+fi
 
 # 3.4. install some casks
 # -------------------------------
@@ -82,40 +119,64 @@ apps=(
   firefox   
   spotify
   dropbox
+  docker
   vlc
   skype
   iterm2
   mactex
   inkscape
   )
-brew cask install --force --appdir="/Applications" ${apps[@]}
-brew cleanup
+
+printf "\nThe following casks will be installed\n"
+printf '   - %s\n' "${taps[@]}"
+printf "\nContinue [y/n] "; read OK
+if [ "$OK" != "Y" ] && [ "$OK" != "y" ]
+then
+  printf "\n Exiting installation...\n"
+  exit 0
+else
+  brew install ${taps[@]}
+  brew cask install --force --appdir="/Applications" ${apps[@]}
+  brew cleanup
+fi
 
 
 # 5. install anaconda
 # ===============================
-wget -nc https://repo.anaconda.com/archive/Anaconda3-5.2.0-MacOSX-x86_64.sh
-sh ./Anaconda3-5.2.0-MacOSX-x86_64.sh -p /usr/local/anaconda
-
-# 5.1 install netcdf and cartopy
-# ------------------------------
-/usr/local/anaconda/bin/conda install netCDF4 cartopy
+printf "\nInstalling Anaconda 5.2? [y/n] "; read OK
+if [ "$OK" != "Y" ] && [ "$OK" != "y" ]
+then
+  printf "\n Exiting installation...\n"
+  exit 0
+else
+  wget -nc -nv https://repo.anaconda.com/archive/Anaconda3-5.2.0-MacOSX-x86_64.sh
+  sh ./Anaconda3-5.2.0-MacOSX-x86_64.sh -s -b -p /usr/local/anaconda
+  /usr/local/anaconda/bin/conda install netCDF4 cartopy
+fi
 
 
 # 7. copy dotfiles
 # ==============================
+
+printf "\nAdd dotfiles [y/n] "; read OK
+if [ "$OK" != "Y" ] && [ "$OK" != "y" ]
+then
+  printf "\n Exiting installation...\n"
+  exit 0
+fi
 
 # git configuration
 git config --global user.name "Daniel Santiago"
 git config --global user.email dspelaez@gmail.com
 
 # clone repository
-git clone ...
+git clone https://github.com/dspelaez/dotfiles.git
 
 # make symbolic links
-cp -r dotfiles $HOME/.dotfiles
+cd dotfiles && cp -r dotfiles $HOME/.dotfiles
 ln -sf $HOME/.dotfiles/latexmk $HOME/.latexmk
 ln -sf $HOME/.dotfiles/matplotlibrc $HOME/.matplotlib/matplotlibrc
+cd ..
 
 
 # 8. vim neovim, tmux conf
@@ -127,6 +188,7 @@ if [ ! -e ~/.config/nvim/autoload/plug.vim ]; then
   curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 fi
+/usr/local/bin/pip install neovim
 ln -sf $HOME/.dotfiles/vimrc $HOME/.vimrc
 ln -sf $HOME/.dotfiles/init.vim $HOME/.config/nvim/init.vim
 nvim +PlugInstall +PlugUpgrade +PlugUpdate +PlugClean! +UpdateRemotePlugins +qall
@@ -136,7 +198,6 @@ nvim +PlugInstall +PlugUpgrade +PlugUpdate +PlugClean! +UpdateRemotePlugins +qal
 # ==============================
 cd ../
 rm -rf tmp
-
 
 
 # --- end of file ---
